Refactor MariaDB installation process and enhance logging
- Introduced spinners for package cache updates and installations to improve user feedback.
- Updated the `installMariaDBPackages` function to accept OS info directly and improved error handling.
- Removed redundant terminal prints and replaced them with debug logs for internal state tracking.
- Modularized discovery functions into separate files for better organization and maintainability.
- Enhanced the `detectDataDirAndSocket` function to handle binary log directory detection.
- Updated default backup directory for configuration to a more appropriate path.
- Improved OS detection to cache results and avoid redundant calls.
- Replaced manual RAM detection with `gopsutil` for better reliability.
- Added new utility functions for MariaDB discovery and configuration management.
