+++ internal/core/mariadb/configure/migration/types.go
@@ -1,0 +1,8 @@
+package migration
+
+type DataMigration struct {
+	Type        string
+	Source      string
+	Destination string
+	Critical    bool
+}
