+++ utils/mariadb/discovery/util.go
@@ -1,0 +1,6 @@
+package mariadb
+
+import "regexp"
+
+// silence unused-import for packages that may be used across files
+var _ = regexp.MustCompile
